requirements.txt
Flask~=3.1.0
flask-cors~=4.0.1
gunicorn>=21.2.0
python-dotenv~=1.0.1
PyJHora==4.8.0
pyswisseph>=2.10.0
geocoder>=1.38.1
timezonefinder>=6.0.0
geopy>=2.4.0
numpy>=1.24.0
pytz>=2024.1
python-dateutil>=2.8.0
Pillow>=10.0.0
